Update the favicon to include a Retina version.
The old favicon is the same as the 16px icon from the Quicksilver app. The updated .ico file includes the 32px icon.

In order to make sure all clients get the right version for both 1x and 2x, this change updates the header template to point at the .ico (and removes the .png file).